decoupled functions into files; updated package-lock.json
