+++ pages/popup.js
@@ -1,3 +1,11 @@
 import React from 'react';
+import { useRouter } from 'next/router';
 
-export default () => <div>Popup</div>;
+const Popup = function() {
+  const router = useRouter();
+  console.log(router.query);
+
+  return <div>POPUP</div>;
+};
+
+export default Popup;
